GEODE-5530 fixes for distributed tests
