+++ geode-core/src/test/java/org/apache/geode/management/internal/cli/commands/CommandAvailabilityIndicatorTest.java
@@ -15,19 +15,12 @@
 
 package org.apache.geode.management.internal.cli.commands;
 
-import static org.assertj.core.api.Assertions.assertThat;
 
-import java.lang.reflect.Method;
-import java.util.List;
 
 import org.junit.Test;
-import org.springframework.shell.core.CommandMarker;
-import org.springframework.shell.core.annotation.CliCommand;
 
-import org.apache.geode.internal.cache.extension.mock.MockExtensionCommands;
-import org.apache.geode.management.cli.CliMetaData;
-import org.apache.geode.management.cli.GfshCommand;
 import org.apache.geode.management.internal.cli.CommandManager;
+
 
 
 public class CommandAvailabilityIndicatorTest {
@@ -35,38 +28,7 @@ public class CommandAvailabilityIndicatorTest {
   @Test
   public void allOnlineCommandsHaveAvailabilityIndicator() {
     CommandManager manager = new CommandManager();
-    assertOnlineCommandsHasAvailabilityIndicator(manager);
+    CommandAvailabilityIndicatorTestHelper.assertOnlineCommandsHasAvailabilityIndicator(manager);
   }
 
-  public static void assertOnlineCommandsHasAvailabilityIndicator(CommandManager manager) {
-    List<CommandMarker> commandMarkers = manager.getCommandMarkers();
-    for (CommandMarker commandMarker : commandMarkers) {
-      // ignore all the other commands beside GfshCommand
-      if (!GfshCommand.class.isAssignableFrom(commandMarker.getClass())) {
-        continue;
-      }
-
-      if (commandMarker instanceof MockExtensionCommands) {
-        continue;
-      }
-
-      for (Method method : commandMarker.getClass().getMethods()) {
-        CliCommand cliCommand = method.getAnnotation(CliCommand.class);
-        if (cliCommand == null) {
-          // the method is not a command method
-          continue;
-        }
-
-        CliMetaData cliMetaData = method.getAnnotation(CliMetaData.class);
-        // all the online commands have availability indicator defined in the commandManager
-        if (cliMetaData == null || cliMetaData != null && !cliMetaData.shellOnly()) {
-          assertThat(manager.getHelper().hasAvailabilityIndicator(cliCommand.value()[0]))
-              .describedAs(cliCommand.value()[0] + " in " + commandMarker.getClass()
-                  + " has no availability indicator defined. "
-                  + "Please add the command in the CommandAvailabilityIndicator")
-              .isTrue();
-        }
-      }
-    }
-  }
 }
