+++ geode-junit/src/main/java/org/apache/geode/internal/cache/extension/mock/DestroyMockCacheExtensionFunction.java
