+++ geode-junit/src/main/java/org/apache/geode/management/internal/cli/dto/ObjectWithCharAttr.java
